+++ src/test/java/model/ladder/HeightTest.java
@@ -1,6 +1,5 @@
-package model;
+package model.ladder;
 
-import model.ladder.Height;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
