+++ src/test/java/model/ladder/LadderTest.java
@@ -1,9 +1,5 @@
-package model;
+package model.ladder;
 
-import model.ladder.Height;
-import model.ladder.Ladder;
-import model.ladder.RandomStatusGenerator;
-import model.ladder.Width;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
