+++ src/test/java/model/player/PlayersTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.player;
 
 import java.util.List;
 import model.players.Players;
